0.3.6+5: update dependencies add another example
