+++ example/android/build.gradle
@@ -1,4 +1,5 @@
 buildscript {
+   
     repositories {
         google()
         jcenter()
@@ -6,7 +7,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.0.0'
+        classpath 'com.android.tools.build:gradle:4.1.1'
     }
 }
