+++ lib/src/geo_point.dart
@@ -1,11 +1,10 @@
+///[GeoPoint]:class contain longitude and latitude of geographic position
 class GeoPoint {
   final double longitude;
   final double latitude;
- 
+
   GeoPoint({
     this.latitude,
     this.longitude,
   });
-
- 
 }
