+++ lib/src/marker.dart
@@ -3,7 +3,13 @@ import 'package:flutter/material.dart';
 class MarkerIcon extends StatelessWidget {
   final Icon icon;
   final AssetImage image;
-  MarkerIcon({this.icon, this.image}) : assert(icon != null || image != null);
+
+  MarkerIcon({
+    this.icon,
+    this.image,
+    Key key,
+  })  : assert(icon != null || image != null),
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
