+++ pubspec.yaml
@@ -9,8 +9,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^1.0.8
-  location_permissions: ^3.0.0
+  flutter_plugin_android_lifecycle: ^1.0.11
+  location_permissions: ^3.0.0+1
   after_layout: ^1.0.7+2
 dev_dependencies:
   flutter_test:
